+++ src/SharpX/FsCheck/ArbitraryIntegerSeq.cs
@@ -1,0 +1,14 @@
+using FsCheck;
+using FsCheck.Fluent;
+
+namespace SharpX.FsCheck;
+
+static class ArbitraryIntegerSeq
+{
+    public static Arbitrary<int[]> Generator()
+    {
+        var seq = Primitives.GenerateSeq<int>(count: 100);
+
+        return Gen.Shuffle(seq).ToArbitrary();
+    }
+}